Sync public branch with internal repository
