+++ src/app/examples/index.ts
@@ -1,18 +1,18 @@
 const examples = Object.freeze([
   {
     name: "Quickstart",
-    path: "quickstart",
+    path: "",
     description:
       "Walkthrough of Smart Account creation, and delegation redemption.",
   },
   {
     name: "Custom Signers",
-    path: "signers",
+    path: "examples/signers",
     description: "See the various signer options in action.",
   },
   {
     name: "Enable and Disable Delegations",
-    path: "toggle-delegation",
+    path: "examples/toggle-delegation",
     description: "Toggle the ability to redeem a delegation.",
   },
 ]);
