+++ src/app/page.tsx
@@ -1,209 +1,307 @@
 "use client";
 
-import { useState } from "react";
-import { createPublicClient, http, zeroAddress } from "viem";
-import { privateKeyToAccount, generatePrivateKey } from "viem/accounts";
-import { sepolia as chain } from "viem/chains";
+import Hero from "@/components/Hero";
 import {
-  Implementation,
-  toMetaMaskSmartAccount,
-  type MetaMaskSmartAccount,
-} from "@codefi/delegator-core-viem";
-import {
-  createBundlerClient,
-  createPaymasterClient,
-  UserOperationReceipt,
-} from "viem/account-abstraction";
-import { createPimlicoClient } from "permissionless/clients/pimlico";
-
-const BUNDLER_URL = process.env.NEXT_PUBLIC_BUNDLER_URL!;
-const RPC_URL = process.env.NEXT_PUBLIC_RPC_URL!;
-// if this is undefined, the API_KEY must be configured to Enable verifying paymaster
-const PAYMASTER_POLICY_ID = process.env.NEXT_PUBLIC_PAYMASTER_POLICY_ID;
-
+  createMetaMaskAccount,
+  createDelegation,
+  executeOnBehalfOfDelegator,
+} from "./quickstart";
 import examples from "@/app/examples";
-import Hero from "../components/Hero";
+import { useState } from "react";
+import {
+  DelegationStruct,
+  Implementation,
+  MetaMaskSmartAccount,
+  getExplorerAddressLink,
+  getExplorerTransactionLink,
+} from "@codefi/delegator-core-viem";
+import { chain, getExplorerUserOperationLink } from "./examples/shared";
+import { UserOperationReceipt } from "viem/account-abstraction";
 
 function App() {
-  const [delegatorSmartAccount, setDelegatorSmartAccount] =
-    useState<MetaMaskSmartAccount<Implementation.Hybrid>>();
-  const [isDeploying, setIsDeploying] = useState(false);
+  const [executeOnBehalfIsLoading, setExecuteOnBehalfIsLoading] =
+    useState(false);
+
+  const [delegatorAccount, setDelegatorAccount] =
+    useState<MetaMaskSmartAccount<Implementation>>();
+  const [delegateAccount, setDelegateAccount] =
+    useState<MetaMaskSmartAccount<Implementation>>();
+  const [delegation, setDelegation] = useState<DelegationStruct>();
   const [userOperationReceipt, setUserOperationReceipt] =
     useState<UserOperationReceipt>();
+  const [userOperationErrorMessage, setUserOperationErrorMessage] =
+    useState<string>();
 
-  const publicClient = createPublicClient({
-    chain,
-    transport: http(RPC_URL),
-  });
+  const [isDelegateDeployed, setIsDelegateDeployed] = useState(false);
+  const [isDelegatorDeployed, setIsDelegatorDeployed] = useState(false);
 
-  const paymasterContext = PAYMASTER_POLICY_ID
-    ? {
-        sponsorshipPolicyId: PAYMASTER_POLICY_ID,
-      }
-    : undefined;
-
-  const pimlicoClient = createPimlicoClient({
-    transport: http(BUNDLER_URL),
-  });
-
-  const paymasterClient = createPaymasterClient({
-    transport: http(BUNDLER_URL),
-  });
-
-  const bundlerClient = createBundlerClient({
-    transport: http(BUNDLER_URL),
-    paymaster: paymasterClient,
-    chain,
-    paymasterContext,
-  });
-
-  async function handleCreateDelegator() {
-    const privateKey = generatePrivateKey();
-    const owner = privateKeyToAccount(privateKey);
-
-    const deploySalt = "0x";
-
-    const smartaccount = await toMetaMaskSmartAccount({
-      client: publicClient,
-      implementation: Implementation.Hybrid,
-      deployParams: [owner.address, [], [], []],
-      deploySalt,
-      signatory: { account: owner },
-    });
-
-    setDelegatorSmartAccount(smartaccount);
-  }
-
-  async function handleDeployDelegator() {
-    setIsDeploying(true);
-    if (!delegatorSmartAccount) {
-      return;
+  const handleCreateDelegator = async () => {
+    try {
+      const account = await createMetaMaskAccount();
+      setDelegatorAccount(account);
+    } catch (error) {
+      console.error(error);
     }
+  };
 
-    // we use a bespoke pimlico client to get the gas price specified by the bundler,
-    // as there is no standard way to do this
-    const { fast: fees } = await pimlicoClient.getUserOperationGasPrice();
+  const handleCreateDelegate = async () => {
+    try {
+      const account = await createMetaMaskAccount();
+      setDelegateAccount(account);
+    } catch (error) {
+      console.error(error);
+    }
+  };
 
-    const gasEstimate = await bundlerClient.estimateUserOperationGas({
-      account: delegatorSmartAccount,
-      calls: [
-        {
-          to: zeroAddress,
-        },
-      ],
-      maxFeePerGas: 1n,
-      maxPriorityFeePerGas: 1n,
-    });
+  const handleCreateDelegation = async () => {
+    if (!delegatorAccount || !delegateAccount) return;
 
-    console.log({ gasEstimate });
+    // Reset downstream state, as it may be a subsequent delegation.
 
-    const userOpHash = await bundlerClient.sendUserOperation({
-      account: delegatorSmartAccount,
-      calls: [
-        {
-          to: zeroAddress,
-        },
-      ],
-      ...fees,
-    });
+    setDelegation(undefined);
+    setUserOperationReceipt(undefined);
 
-    const receipt = await bundlerClient.waitForUserOperationReceipt({
-      hash: userOpHash,
-    });
+    try {
+      const delegation = await createDelegation(
+        delegatorAccount,
+        delegateAccount.address
+      );
+      setDelegation(delegation);
+    } catch (error) {
+      console.error(error);
+    }
+  };
 
-    setIsDeploying(false);
-    setUserOperationReceipt(receipt);
-  }
+  const handleExecuteOnBehalf = async () => {
+    if (!delegateAccount || !delegatorAccount || !delegation) return;
+
+    setUserOperationReceipt(undefined);
+
+    setExecuteOnBehalfIsLoading(true);
+
+    const { factory, factoryData } = await delegatorAccount.getFactoryArgs();
+
+    const factoryArgs =
+      factory && factoryData ? { factory, factoryData } : undefined;
+
+    try {
+      const receipt = await executeOnBehalfOfDelegator(
+        delegateAccount,
+        delegation,
+        factoryArgs
+      );
+      if (receipt.success) {
+        setUserOperationReceipt(receipt);
+      } else {
+        throw new Error(`User operation failed: ${receipt.reason}`);
+      }
+    } catch (error) {
+      setUserOperationErrorMessage((error as Error).message);
+    }
+    setExecuteOnBehalfIsLoading(false);
+
+    delegateAccount.isDeployed().then(setIsDelegateDeployed);
+    delegatorAccount.isDeployed().then(setIsDelegatorDeployed);
+  };
+
+  const handleStartAgain = () => {
+    setDelegatorAccount(undefined);
+    setDelegateAccount(undefined);
+    setDelegation(undefined);
+    setUserOperationReceipt(undefined);
+  };
 
   return (
     <div className="mx-auto">
       <Hero />
-      <h2 className="text-2xl font-bold mb-4">Overview</h2>
-      <p className="mb-4">
-        A basic template to get you started with the{" "}
-        <a
-          className="text-green-500"
-          href="https://metamask.io/developer/delegation-toolkit/"
-          target="_blank"
-        >
-          Delegation Toolkit
-        </a>
-        . It includes a collection of self contained examples (selectable via
-        the dropdown in the top-right or via the cards below) which can simply
-        be deleted if you're using this as a scaffold for a new project.
-      </p>
       <h2 className="text-2xl font-bold mb-4">
         Hello <span className="line-through">World</span>{" "}
         <span className="bg-clip-text text-transparent bg-gradient-to-b from-white to-green-500">
           Gator
-        </span>
+        </span>{" "}
+        - Quickstart
       </h2>
       <p className="mb-4">
-        Click the following to create a new Delegator account.
+        This example demonstrates how to get started with the Delegation
+        Toolkit. See the{" "}
+        <a
+          href="https://docs.gator.metamask.io/get-started/quickstart"
+          className="text-green-500"
+          target="_blank"
+        >
+          accompanying documentation
+        </a>
+        .
       </p>
-      <button
-        className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
-        onClick={handleCreateDelegator}
-        disabled={!!delegatorSmartAccount}
-      >
-        Create Delegator Account
-      </button>
-      <p className="mb-4" style={{ overflow: "auto" }}>
-        {delegatorSmartAccount ? (
-          <a
-            href={`https://sepolia.etherscan.io/address/${delegatorSmartAccount.address}`}
-            className="text-green-500"
-            target="_blank"
-          >
-            {userOperationReceipt && (
-              <span className="inline-block animate-ping mr-2">🐊</span>
-            )}
-            {delegatorSmartAccount.address}
-            {userOperationReceipt && (
-              <span className="inline-block animate-ping ml-2">🐊</span>
-            )}
-          </a>
-        ) : (
-          "NA"
-        )}
-        {isDeploying && (
-          <span>
-            {" "}
-            <span className="inline-block animate-spin">🐊</span> Deploying{" "}
-            <span className="inline-block animate-spin">🐊</span>
-          </span>
-        )}
-      </p>
-      {delegatorSmartAccount && !userOperationReceipt && (
+      <div className="space-y-4">
         <div>
-          <p className="mb-4">
-            Nice! You've just created a counterfactual (meaning it's not yet
-            deployed on chain) Delegator account. Click below to deploy it via a
-            pre-configured bundler.
+          <h3 className="text-lg font-bold">Delegator Account</h3>
+          <p>
+            The MetaMask smart contract account that grants authority. Initially
+            this will be counterfactual (not deployed on-chain), and will be
+            deployed in the same user operation, just in time for redeeming the
+            delegation.
+          </p>
+          {!delegatorAccount && (
+            <button
+              className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
+              onClick={handleCreateDelegator}
+            >
+              Create Delegator Account
+            </button>
+          )}
+          {delegatorAccount && (
+            <div>
+              <a
+                href={getExplorerAddressLink(
+                  chain.id,
+                  delegatorAccount.address
+                )}
+                target="_blank"
+                className="text-green-500 font-mono"
+              >
+                {delegatorAccount.address}
+              </a>{" "}
+              - {isDelegatorDeployed ? "Deployed" : "Counterfactual"}
+            </div>
+          )}
+        </div>
+        <div>
+          <h3 className="text-lg font-bold">Delegate Account</h3>
+          <p>
+            The MetaMask smart contract account that receives the{" "}
+            <span className="font-mono">delegation</span>. Initially this will
+            be counterfactual (not deployed on-chain), until it is deployed by
+            submitting a user operation.
+          </p>
+          {!delegateAccount && (
+            <button
+              className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
+              onClick={handleCreateDelegate}
+            >
+              Create Delegate Account
+            </button>
+          )}
+
+          {delegateAccount && (
+            <div>
+              <a
+                href={getExplorerAddressLink(chain.id, delegateAccount.address)}
+                target="_blank"
+                className="text-green-500 font-mono"
+              >
+                {delegateAccount.address}
+              </a>{" "}
+              - {isDelegateDeployed ? "Deployed" : "Counterfactual"}
+            </div>
+          )}
+        </div>
+
+        <div>
+          <h3 className="text-lg font-bold">Delegation</h3>
+          <p>
+            The <span className="font-mono">delegator</span> creates and signs a{" "}
+            <span className="font-mono">delegation</span>, granting specific
+            authority to the <span className="font-mono">delegate account</span>
+            . In this case, the delegation allows a transfer of 0 ether to the
+            zero address.
+          </p>
+
+          <button
+            className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
+            onClick={handleCreateDelegation}
+            disabled={
+              !delegatorAccount || !delegateAccount || executeOnBehalfIsLoading
+            }
+          >
+            Create Delegation
+          </button>
+
+          {delegation && (
+            <div className="mt-2 p-2 bg-gray-800 rounded">
+              <pre className="whitespace-pre-wrap break-all">
+                {JSON.stringify(
+                  delegation,
+                  (_, v) => (typeof v === "bigint" ? `${v.toString()}n` : v),
+                  2
+                )}
+              </pre>
+            </div>
+          )}
+        </div>
+
+        <div>
+          <h3 className="text-lg font-bold">Execute</h3>
+          <p>
+            The redeemer submits a user operation that executes the action
+            allowed by the <span className="font-mono">delegation</span> (in
+            this case, transfer nothing to no one) on behalf of the{" "}
+            <span className="font-mono">delegator</span>. If the{" "}
+            <span className="font-mono">delegator</span> is counterfactual, it
+            will be deployed as a separate{" "}
+            <span className="font-mono">Call</span> in the same user operation.
           </p>
           <button
             className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
-            onClick={handleDeployDelegator}
-            disabled={isDeploying || !delegatorSmartAccount}
+            onClick={handleExecuteOnBehalf}
+            disabled={!delegation || executeOnBehalfIsLoading}
           >
-            Deploy Delegator Account
+            Execute
           </button>
+
+          {executeOnBehalfIsLoading && (
+            <span className="animate-spin inline-block ml-2">
+              🐊 loading...
+            </span>
+          )}
+          {userOperationReceipt && (
+            <div>
+              User operation hash:{" "}
+              <a
+                href={getExplorerUserOperationLink(
+                  chain.id,
+                  userOperationReceipt.userOpHash
+                )}
+                className="text-green-500 font-mono"
+                target="_blank"
+              >
+                {userOperationReceipt.userOpHash}
+              </a>
+              <br />
+              Transaction hash:{" "}
+              <a
+                href={getExplorerTransactionLink(
+                  chain.id,
+                  userOperationReceipt.receipt.transactionHash
+                )}
+                className="text-green-500 font-mono"
+                target="_blank"
+              >
+                {userOperationReceipt.receipt.transactionHash}
+              </a>
+            </div>
+          )}
+          {userOperationErrorMessage && (
+            <div className="mt-2 p-2 bg-gray-800 rounded">
+              <pre className="whitespace-pre-wrap break-all">
+                Error submitting User Operation: {userOperationErrorMessage}
+              </pre>
+            </div>
+          )}
+          <div className="mt-4">
+            <button
+              onClick={handleStartAgain}
+              disabled={
+                (!delegateAccount && !delegatorAccount) ||
+                executeOnBehalfIsLoading
+              }
+              className="bg-white text-black rounded-md px-2 py-1 disabled:opacity-50 disabled:cursor-not-allowed hover:bg-gray-200"
+            >
+              Start over
+            </button>
+          </div>
         </div>
-      )}
-      {userOperationReceipt && (
-        <p className="mb-4">
-          <a
-            className="text-green-500"
-            href={`https://sepolia.etherscan.io/tx/${userOperationReceipt.receipt.transactionHash}`}
-            target="_blank"
-          >
-            Done!{" "}
-          </a>
-          Click on the address above to check it out. Now feel free to try out
-          some of the examples below to see what else you can do with the
-          Delegation Toolkit.
-        </p>
-      )}
+      </div>
 
       <h2 className="text-2xl font-bold mb-4">Examples</h2>
       <div className="flex flex-wrap gap-4 mb-4">
